Add sqlite-based blockchain storage
This blockchain storage is implemented using a series of SQL tables and views.
(Oh the irony of using old-school tech like SQL to do blockchains!)

It contains tables for blocks, transaction, tx inputs, outputs, and
transaction–block relationship. It creates various views such as one for
viewing full transaction information, one for checking unspent transaction
outputs, one for unauthorized spending (spending money not belonging to the
spender), overspent transactions, etc.

Currently the blocks and transactions use their hash as the primary key. This
is probably not very efficient. The transaction_full view also isn't very
inefficient and is not used.