Refactor client code to use classes
The main problems with the previous functional approach are:

(a) scope; we have lots of variable shadowing, as well as `nonlocal`
(b) reduce nesting when we have many context managers and async context
    managers

The idea of subclassing `AsyncExitStack` was discovered by reading the Stack
Overflow answer for composing context managers. It is surprisingly difficult
to find information, but see https://stackoverflow.com/q/51649361/1102119 for
details. The async variant of the context manager isn't documented to have the
sync methods too, but reading through the original discussion in the Python
bug tracker https://bugs.python.org/issue29302 and the GitHub PR
https://github.com/python/cpython/pull/4790/files shows that this is the case.