Handle Control-C more gracefully
Instead of raising KeyboardInterrupt in the main thread, terminating the
network connection but leaving the readline() running, we now catch them and
print a message.

In this program, the only correct way to quit is through EOF, which is
reported by readline(), which exits the `handle_user_interaction` function and
then cancels all threads.