Distinguish between all and mineable tentative transactions
We pessimistically assume that we only want to mine transactions directly
descended from the current longest chain, not any more descendants, because
the validity of those transactions could depend on other unmined transactions,
which could then be mined by someone before us.

This also helps when there are minor forks.