Consistency check now happens with a view of current ancestors
This means that every time we receive a block, certains checks that refer to
other blocks will happen with a view where the only data are from the current
chain. This includes two checks: input referring to a nonexistent output, and
double spent. Other checks that do not require blocks (such as the
unauthorized spending check and overspending check) remain the same.