Again fix transaction selection when mining
One issue is that previously, we are still seeing mining tasks fail because
it's trying to select a transaction that has already been selected, leading to
unique constraint failures. This is fixed by making the
all_tentative_txns exclude these transactions and making the provisional block
the longest.

Another issue is that the coinbase transaction of another peer is included in
the mined block. This is also removed.