Partially handle orphan blocks
Let's say while we are synchronizing, someone else mined a new block that
makes it the new longest block:

                 _ o __ o __ o     <- someone else longest
                /
    o -- o -- o -- o -- o -- o     <- server longest

becomes

                 _ o __ o __ o __ o     <- the longest
                /
    o -- o -- o -- o -- o -- o     <- server abandoned

Then we would receive the announcement for the newly mined block but we cannot
process it.

The complete solution is to support orphaned blocks, but I'm too lazy to do
that. So a partial solution is that whenever a receive failed, we set a flag
to do a sync after processing this message. Afterwards, we actually do the
sync, hopefully we will re-receive the failed block/txn and process it
correctly. Note that for blocks that failed, the reason is probably that a
parent block is missing, but we will then receive those missing parent blocks;
for txns that failed, the reason is probably that a parent transaction is in a
block that has been mined between the time we got the longest chain and the
tentative transactions, but we will then receive the new longest chain.