Change utxo to not use global transactions
This is rather like a continuation of the work in ca82e14, but for UTXO.
Previously we assume any unspent transaction output is valid, but that is not
really the case with side chains etc. Instead, UTXO is deemed valid using some
rules similar to that in the Bitcoin client:

https://github.com/bitcoin/bitcoin/blob/3515612e069e3730f173a08e60b99f96b174188d/src/wallet/wallet.h#L571-L575

We treat a UTXO as valid if it is either in a mined block, or comes from a
trusted wallet and is not a block reward. Originally I tried to make a
transaction valid if either it is in a mined block or it has parents and all
of its parents are valid. It turns out this recursion is too hard to express
in SQL. Also it is rather too liberal.

Note that the UTXO is required to be unspent globally, not just unspent
subject to those conditions.

The notion of trustworthy wallet is admittedly not that elegant, but is
pragmatic.